requirements.txt
requests
sqlalchemy
pyyaml